# frontend/docs-frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MyHouse API",
    "description": "Real estate rental matching platform (owners ↔ seekers).",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "The access token itself keeps authenticating requests until its own\n(max 15 min) expiry — MH-35's stateless JWTs can't be invalidated early.\nLogout only guarantees the session can't be *refreshed* again.",
        "operationId": "logout",
        "responses": {
          "200": {
            "description": "Current refresh token revoked; cookie cleared"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        }
      }
    },
    "/api/v1/auth/otp/request": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Identical 200 response whether `payload.email` is registered or not —\nthe endpoint must never let a caller enumerate accounts.",
        "operationId": "otp_request",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OTP code generated and emailed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpRequestResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests for this email within the rate-limit window"
          }
        }
      }
    },
    "/api/v1/auth/otp/verify": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Second half of the unified auth endpoint (seeker_auth_flow.mermaid §2):\nverifies the code cached by `/auth/otp/request`, creates the account on\nfirst login, and issues the session.",
        "operationId": "otp_verify",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpVerifyDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OTP verified; session issued (account created if new)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpVerifyResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid, expired, or attempt-exhausted OTP code"
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "refresh",
        "responses": {
          "200": {
            "description": "New access token issued; refresh cookie rotated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, unknown, expired, or reused refresh token"
          }
        }
      }
    },
    "/api/v1/listings": {
      "get": {
        "tags": [
          "listings"
        ],
        "summary": "Public paginated feed of listings (cover photo + summary).",
        "operationId": "list",
        "parameters": [
          {
            "name": "owner_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "city",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ListingType"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of listings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_ListingSummaryDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/listings/{id}": {
      "get": {
        "tags": [
          "listings"
        ],
        "summary": "Full listing detail — description, media, owner info (no phone).",
        "operationId": "get_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Listing id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Listing detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Listing not found"
          }
        }
      }
    },
    "/api/v1/users/me": {
      "put": {
        "tags": [
          "users"
        ],
        "operationId": "update_me",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Profile updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing or oversized first_name, last_name or phone"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Checks PostgreSQL connectivity via `SELECT 1`.",
        "operationId": "check",
        "responses": {
          "200": {
            "description": "Database reachable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthStatus"
                }
              }
            }
          },
          "503": {
            "description": "Database unreachable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthStatus"
                }
              }
            }
          }
        }
      }
    },
    "/health/storage": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Throwaway smoke check confirming `Arc<dyn StorageProvider>` (MH-27) is reachable\nfrom a handler via `State<AppState>` extraction. No module has real uploads yet\n(first real caller is EP-08/Media) — this only proves the plumbing compiles and\nresolves through the extractor, mirroring [`check`] above.",
        "operationId": "check_storage",
        "responses": {
          "200": {
            "description": "Storage provider reachable via AppState",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageStatus"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HealthStatus": {
        "type": "object",
        "description": "Health check response body.",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string"
          }
        }
      },
      "ListingDetailDto": {
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "type",
          "status",
          "city",
          "price",
          "media",
          "owner",
          "created_at"
        ],
        "properties": {
          "city": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "media": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ListingMediaDto"
            }
          },
          "neighborhood": {
            "type": [
              "string",
              "null"
            ]
          },
          "owner": {
            "$ref": "#/components/schemas/OwnerDetailDto"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "rooms": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "status": {
            "$ref": "#/components/schemas/ListingStatus"
          },
          "surface_m2": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/ListingType"
          }
        }
      },
      "ListingDetailResponse": {
        "type": "object",
        "description": "Envelope for `GET /listings/:id` — single-object `{ \"data\": {...} }`,\ndistinct from the paginated list envelope used by `GET /listings`.",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ListingDetailDto"
          }
        }
      },
      "ListingMediaDto": {
        "type": "object",
        "description": "One media attachment in a listing detail response.",
        "required": [
          "id",
          "url",
          "is_cover",
          "position"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "is_cover": {
            "type": "boolean"
          },
          "position": {
            "type": "integer",
            "format": "int32"
          },
          "url": {
            "type": "string"
          }
        }
      },
      "ListingStatus": {
        "type": "string",
        "enum": [
          "available",
          "unavailable"
        ]
      },
      "ListingSummaryDto": {
        "type": "object",
        "required": [
          "id",
          "title",
          "type",
          "status",
          "city",
          "price",
          "owner"
        ],
        "properties": {
          "city": {
            "type": "string"
          },
          "cover_photo_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "neighborhood": {
            "type": [
              "string",
              "null"
            ]
          },
          "owner": {
            "$ref": "#/components/schemas/OwnerSummaryDto"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "$ref": "#/components/schemas/ListingStatus"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/ListingType"
          }
        }
      },
      "ListingType": {
        "type": "string",
        "enum": [
          "apartment",
          "studio",
          "house",
          "room",
          "villa",
          "other"
        ]
      },
      "OtpRequestDto": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string"
          }
        }
      },
      "OtpRequestMessageDto": {
        "type": "object",
        "description": "`{ \"data\": { \"message\": \"...\" } }` — identical shape whether `email` is\nknown or unknown, so the response can never be used to enumerate\nregistered accounts.",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "OtpRequestResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/OtpRequestMessageDto"
          }
        }
      },
      "OtpVerifyDto": {
        "type": "object",
        "required": [
          "email",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
      "OtpVerifyResponse": {
        "type": "object",
        "description": "`{ \"data\": { \"access_token\": \"...\", \"is_new_user\": bool } }` per\n`TECHNICAL_SPEC_MVP.md §4.1`.",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/OtpVerifyTokenDto"
          }
        }
      },
      "OtpVerifyTokenDto": {
        "type": "object",
        "required": [
          "access_token",
          "is_new_user"
        ],
        "properties": {
          "access_token": {
            "type": "string"
          },
          "is_new_user": {
            "type": "boolean"
          }
        }
      },
      "OwnerDetailDto": {
        "type": "object",
        "description": "Owner info as embedded in a listing detail (adds avatar, still no phone —\nphone reveal is gated behind `GET /listings/:id/contact`, out of scope here).",
        "required": [
          "id"
        ],
        "properties": {
          "avatar_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "first_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "last_name": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "OwnerSummaryDto": {
        "type": "object",
        "description": "Owner info as embedded in a listing summary (no avatar, no phone).",
        "required": [
          "id"
        ],
        "properties": {
          "first_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "last_name": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "PaginatedResponse_ListingSummaryDto": {
        "type": "object",
        "description": "Full paginated response envelope.\n\nGeneric over the item type so every module can reuse it:\n`PaginatedResponse<ListingSummaryDto>`, `PaginatedResponse<UserDto>`, etc.",
        "required": [
          "data",
          "pagination"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "title",
                "type",
                "status",
                "city",
                "price",
                "owner"
              ],
              "properties": {
                "city": {
                  "type": "string"
                },
                "cover_photo_url": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "neighborhood": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "owner": {
                  "$ref": "#/components/schemas/OwnerSummaryDto"
                },
                "price": {
                  "type": "number",
                  "format": "double"
                },
                "status": {
                  "$ref": "#/components/schemas/ListingStatus"
                },
                "title": {
                  "type": "string"
                },
                "type": {
                  "$ref": "#/components/schemas/ListingType"
                }
              }
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        }
      },
      "PaginationMeta": {
        "type": "object",
        "description": "Pagination metadata included alongside `data` in list responses.",
        "required": [
          "page",
          "per_page",
          "total",
          "total_pages"
        ],
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "per_page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "total_pages": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
      "RefreshResponse": {
        "type": "object",
        "description": "`{ \"data\": { \"access_token\": \"...\" } }` — matches the envelope used\nelsewhere (e.g. `ListingDetailResponse`) and by the sibling\n`/auth/otp/verify` endpoint's actual JSON example.",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RefreshTokenDto"
          }
        }
      },
      "RefreshTokenDto": {
        "type": "object",
        "required": [
          "access_token"
        ],
        "properties": {
          "access_token": {
            "type": "string"
          }
        }
      },
      "Role": {
        "type": "string",
        "enum": [
          "seeker",
          "owner",
          "admin"
        ]
      },
      "StorageStatus": {
        "type": "object",
        "description": "Storage backend response body.",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string"
          }
        }
      },
      "UpdateMeDto": {
        "type": "object",
        "required": [
          "first_name",
          "last_name"
        ],
        "properties": {
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "phone": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "UserDto": {
        "type": "object",
        "required": [
          "id",
          "email",
          "role"
        ],
        "properties": {
          "avatar_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": "string"
          },
          "first_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "last_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "phone": {
            "type": [
              "string",
              "null"
            ]
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UserDto"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "health",
      "description": "Service health check"
    },
    {
      "name": "listings",
      "description": "Public read-only listings feed"
    },
    {
      "name": "auth",
      "description": "Authentication — OTP request, token refresh, logout"
    },
    {
      "name": "users",
      "description": "Authenticated user profile"
    }
  ]
}